move more utility functions
