+++ bigraph_schema/__init__.py
@@ -2,6 +2,7 @@ from bigraph_schema.registry import (
     deep_merge, validate_merge, default, Registry, hierarchy_depth, is_schema_key,
     strip_schema_keys, type_parameter_key, non_schema_keys, set_path, transform_path)
 from bigraph_schema.type_functions import (
-    get_path, establish_path, visit_method)
+    establish_path, visit_method)
+from bigraph_schema.utilities import get_path
 from bigraph_schema.edge import Edge
 from bigraph_schema.type_system import TypeSystem
