+++ bigraph_schema/type_functions.py
@@ -82,7 +82,8 @@ from bigraph_schema.registry import (
     is_schema_key, non_schema_keys, type_parameter_key,
     deep_merge, hierarchy_depth,
     establish_path)
-from bigraph_schema.utilities import is_empty, union_keys, tuple_from_type, array_shape, read_datatype, read_shape
+from bigraph_schema.utilities import is_empty, union_keys, tuple_from_type, array_shape, read_datatype, read_shape, \
+    remove_path, type_parameters_for
 
 # Create a new module dynamically for the dataclasses
 module_name = 'bigraph_schema.data'
@@ -116,25 +117,6 @@ merge_schema_keys = (
     '_ports',
     '_type_parameters',
 )
-
-def diff(a, b):
-    if isinstance(a, dict) and isinstance(b, dict):
-        result = {}
-        for key in union_keys(a, b):
-            if key in a:
-                if key in b:
-                    inner = diff(a[key], b[key])
-                    if inner:
-                        result[key] = inner
-                else:
-                    result[key] = f'A: {a[key]}\nB: (missing)'
-            else:
-                result[key] = f'A: (missing)\nB: {b[key]}'
-        if result:
-            return result
-    else:
-        if a != b:
-            return f'A: {a}\nB: {b}'
 
 
 def type_merge(dct, merge_dct, path=tuple(), merge_supers=False):
@@ -171,39 +153,6 @@ def type_merge(dct, merge_dct, path=tuple(), merge_supers=False):
             
     return dct
 
-def get_path(tree, path):
-    """
-    Given a tree and a path, find the subtree at that path
-
-    Args:
-    - tree: the tree we are looking in (a nested dict)
-    - path: a list/tuple of keys we follow down the tree to find the subtree we are looking for
-
-    Returns:
-    - subtree: the subtree found by following the list of keys down the tree
-    """
-
-    if len(path) == 0:
-        return tree
-    else:
-        head = path[0]
-        if not tree or head not in tree:
-            return None
-        else:
-            return get_path(tree[head], path[1:])
-
-def remove_path(tree, path):
-    """
-    Removes whatever subtree lives at the given path
-    """
-
-    if path is None or len(path) == 0:
-        return None
-
-    upon = get_path(tree, path[:-1])
-    if upon is not None:
-        del upon[path[-1]]
-    return tree
 
 def visit_method(schema, state, method, values, core):
     """
@@ -233,14 +182,6 @@ def visit_method(schema, state, method, values, core):
         core)
 
     return result
-
-def type_parameters_for(schema):
-    parameters = []
-    for key in schema['_type_parameters']:
-        subschema = schema.get(f'_{key}', 'any')
-        parameters.append(subschema)
-
-    return parameters
 
 
 # =========================
